Use a more inuitive data type for ordering
