+++ src/routes/board/+page.server.ts
@@ -1,5 +1,5 @@
 import prisma from "$lib/utils/prisma";
-import { sortBoardPos } from "$lib/utils/sortCommittee";
+import { sortBoardPos } from "$lib/utils/committee-ordering/sort";
 import type { PageServerLoad } from "./$types";
 
 export const load: PageServerLoad = async () => {
